+++ test/test_agent.py
@@ -1,57 +1,189 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+""" Agent Class
+
+__author__: Conor Heins, Alexander Tschantz, Daphne Demekas, Brennan Klein
+
+"""
+
 import os
 import unittest
 
 import numpy as np
-from scipy.io import loadmat
 
-from pymdp.agent import Agent, build_belief_array, build_xn_vn_array
-from pymdp.core.utils import random_A_matrix, random_B_matrix, obj_array_zeros, get_model_dimensions, convert_observation_array
-from pymdp.core.utils import to_arr_of_arr, to_numpy
-from pymdp.core import control
-
-DATA_PATH = "test/matlab_crossval/output/"
+from pymdp.agent import Agent
+from pymdp.core import utils
+from pymdp.core import inference, control
 
 class TestAgent(unittest.TestCase):
    
-    def test_fpi_inference(self):
+    def test_agent_init_with_control_fac_idx(self):
+        """
+        Initialize instance of the agent class and pass in a custom `control_fac_idx`
+        """
+
         num_obs = [2, 4]
         num_states = [2, 2]
-        num_control = [2, 2]
-        A = random_A_matrix(num_obs, num_states)
-        B = random_B_matrix(num_states, num_control)
+        num_controls = [2, 2]
+        A = utils.random_A_matrix(num_obs, num_states)
+        B = utils.random_B_matrix(num_states, num_controls)
 
-        C = obj_array_zeros([num_ob for num_ob in num_obs])
-        C[1][0] = 1.0  
-        C[1][1] = -2.0  
+        agent = Agent(A=A, B=B, control_fac_idx=[1])
 
-        agent = Agent(A=A, B=B, C=C, control_fac_idx=[1])
-        o, s = [0, 2], [0, 0]
-        qx = agent.infer_states(o)
-        agent.infer_policies()
-        action = agent.sample_action()
+        self.assertEqual(agent.num_controls[0], 1)
+        self.assertEqual(agent.num_controls[1], 2)
+    
+    def test_agent_init_without_control_fac_idx(self):
+        """
+        Initialize instance of the agent class and pass in a custom `control_fac_idx`
+        """
 
-        self.assertEqual(len(action), len(num_control))
-
-    def test_mmp_inference(self):
         num_obs = [2, 4]
         num_states = [2, 2]
-        num_control = [2, 2]
-        A = random_A_matrix(num_obs, num_states)
-        B = random_B_matrix(num_states, num_control)
+        num_controls = [2, 2]
+        A = utils.random_A_matrix(num_obs, num_states)
+        B = utils.random_B_matrix(num_states, num_controls)
 
-        C = obj_array_zeros(num_obs)
-        C[1][0] = 1.0  
-        C[1][1] = -2.0  
+        agent = Agent(A=A, B=B)
 
-        planning_horizon = 5
+        self.assertEqual(agent.num_controls[0], 2)
+        self.assertEqual(agent.num_controls[1], 2)
+
+        self.assertEqual([0, 1], agent.control_fac_idx)
+
+    def test_reset_agent_VANILLA(self):
+        """
+        Ensure the `reset` method of Agent() using the new refactor is working as intended, 
+        using the `VANILLA` argument to `inference_algo` 
+        """
+
+        num_obs = [2, 4]
+        num_states = [2, 3]
+        num_controls = [2, 3]
+        A = utils.random_A_matrix(num_obs, num_states)
+        B = utils.random_B_matrix(num_states, num_controls)
+
+        agent = Agent(A=A, B=B, inference_algo = "VANILLA")
+
+        init_qs = utils.obj_array_uniform(agent.num_states)
+        self.assertTrue(all( [ (agent.qs[f] == init_qs[f]).all() for f in range(agent.num_factors)]) )
+        self.assertTrue(agent.curr_timestep == 0)
+
+    def test_reset_agent_MMP_wBMA(self):
+        """
+        Ensure the `reset` method of Agent() using the new refactor is working as intended, 
+        using the `MMP` argument to `inference_algo`, and `use_BMA` equal to True
+        """
+
+        num_obs = [2, 4]
+        num_states = [2, 3]
+        num_controls = [2, 3]
+        A = utils.random_A_matrix(num_obs, num_states)
+        B = utils.random_B_matrix(num_states, num_controls)
+
+        agent = Agent(A=A, B=B, inference_algo = "MMP", use_BMA=True)
+
+        self.assertEqual(len(agent.latest_belief), agent.num_factors)
+        self.assertTrue(all( [ (agent.latest_belief[f] == agent.D[f]).all() for f in range(agent.num_factors)]) )
+        self.assertTrue(agent.curr_timestep == 0)
+    
+    def test_reset_agent_MMP_wPSP(self):
+        """
+        Ensure the `reset` method of Agent() using the new refactor is working as intended, 
+        using the `MMP` argument to `inference_algo`, and `policy-separated prior` equal to True
+        """
+
+        num_obs = [2, 4]
+        num_states = [2, 3]
+        num_controls = [2, 3]
+        A = utils.random_A_matrix(num_obs, num_states)
+        B = utils.random_B_matrix(num_states, num_controls)
+
+        agent = Agent(A=A, B=B, inference_algo = "MMP", use_BMA = False, policy_sep_prior = True)
+
+        self.assertEqual(len(agent.qs[0]) - agent.inference_horizon - 1, agent.policy_len)
+    
+    def test_agent_infer_states(self):
+        """
+        Test `infer_states` method of the Agent() class
+        """
+
+        ''' VANILLA method (fixed point iteration) with one hidden state factor and one observation modality '''
+        num_obs = [5]
+        num_states = [3]
+        num_controls = [1]
+        A = utils.random_A_matrix(num_obs, num_states)
+        B = utils.random_B_matrix(num_states, num_controls)
+
+        agent = Agent(A=A, B=B, inference_algo = "VANILLA")
+
+        o = tuple([np.random.randint(obs_dim) for obs_dim in num_obs])
+        qs_out = agent.infer_states(o)
+
+        qs_validation = inference.update_posterior_states(A, o, prior=agent.D)
+
+        for f in range(len(num_states)):
+            self.assertTrue(np.isclose(qs_validation[f], qs_out[f]).all())
+
+        ''' VANILLA method (fixed point iteration) with multiple hidden state factors and multiple observation modalities '''
+        num_obs = [2, 4]
+        num_states = [2, 3]
+        num_controls = [2, 3]
+        A = utils.random_A_matrix(num_obs, num_states)
+        B = utils.random_B_matrix(num_states, num_controls)
+
+        agent = Agent(A=A, B=B, inference_algo = "VANILLA")
+
+        o = tuple([np.random.randint(obs_dim) for obs_dim in num_obs])
+        qs_out = agent.infer_states(o)
+
+        qs_validation = inference.update_posterior_states(A, o, prior=agent.D)
+
+        for f in range(len(num_states)):
+            self.assertTrue(np.isclose(qs_validation[f], qs_out[f]).all())
+
+        ''' Marginal message passing inference with multiple hidden state factors and multiple observation modalities '''
+        num_obs = [5]
+        num_states = [3]
+        num_controls = [1]
+        A = utils.random_A_matrix(num_obs, num_states)
+        B = utils.random_B_matrix(num_states, num_controls)
+
+        agent = Agent(A=A, B=B, inference_algo = "MMP")
+
+        o = tuple([np.random.randint(obs_dim) for obs_dim in num_obs])
+        qs_pi_out = agent.infer_states(o)
+
+        policies = control.construct_policies(num_states, num_controls, policy_len = 1)
+
+        qs_pi_validation, _ = inference.update_posterior_states_v2(A, B, [o], policies, prior = agent.D, policy_sep_prior = False)
+
+        for p_idx in range(len(policies)):
+            for f in range(len(num_states)):
+                self.assertTrue(np.isclose(qs_pi_validation[p_idx][0][f], qs_pi_out[p_idx][0][f]).all())
+
+        ''' Marginal message passing inference with multiple hidden state factors and multiple observation modalities '''
+        num_obs = [2, 4]
+        num_states = [2, 2]
+        num_controls = [2, 2]
+        A = utils.random_A_matrix(num_obs, num_states)
+        B = utils.random_B_matrix(num_states, num_controls) 
+
+        planning_horizon = 3
         backwards_horizon = 1
-        agent = Agent(A=A, B=B, C=C, control_fac_idx=[1], inference_algo="MMP", policy_len=planning_horizon, inference_horizon=backwards_horizon)
+        agent = Agent(A=A, B=B, inference_algo="MMP", policy_len=planning_horizon, inference_horizon=backwards_horizon)
         o = [0, 2]
-        qs_pi = agent.infer_states(o)
+        qs_pi_out = agent.infer_states(o)
 
-        for qs_pi_i in qs_pi:
-            self.assertEqual(len(qs_pi_i), planning_horizon + backwards_horizon)
+        policies = control.construct_policies(num_states, num_controls, policy_len = planning_horizon)
 
+        qs_pi_validation, _ = inference.update_posterior_states_v2(A, B, [o], policies, prior = agent.D, policy_sep_prior = False)
+
+        for p_idx in range(len(policies)):
+            for t in range(planning_horizon+backwards_horizon):
+                for f in range(len(num_states)):
+                    self.assertTrue(np.isclose(qs_pi_validation[p_idx][t][f], qs_pi_out[p_idx][t][f]).all())
 
     def test_mmp_active_inference(self):
         """
@@ -61,11 +193,11 @@ class TestAgent(unittest.TestCase):
 
         num_obs = [3, 2]
         num_states = [4, 3]
-        num_control = [1, 3]
-        A = random_A_matrix(num_obs, num_states)
-        B = random_B_matrix(num_states, num_control)
+        num_controls = [1, 3]
+        A = utils.random_A_matrix(num_obs, num_states)
+        B = utils.random_B_matrix(num_states, num_controls)
 
-        C = obj_array_zeros(num_obs)
+        C = utils.obj_array_zeros(num_obs)
         C[1][0] = 1.0  
         C[1][1] = -2.0  
 
@@ -80,72 +212,10 @@ class TestAgent(unittest.TestCase):
             agent.infer_policies()
             action = agent.sample_action()
         
-        print(agent.prev_actions)
-        print(agent.prev_obs)
-    
-    def test_active_inference_SPM_1a(self):
-        """
-        Test against output of SPM_MDP_VB_X.m
-        1A - one hidden state factor, one observation modality, backwards horizon = 3, policy_len = 1, policy-conditional prior
-        """
-        array_path = os.path.join(os.getcwd(), DATA_PATH + "vbx_test_1a.mat")
-        mat_contents = loadmat(file_name=array_path)
-
-        A = mat_contents["A"][0]
-        B = mat_contents["B"][0]
-        C = to_arr_of_arr(mat_contents["C"][0][0][:,0])
-        obs_matlab = mat_contents["obs"].astype("int64")
-        policy = mat_contents["policies"].astype("int64") - 1
-        t_horizon = mat_contents["t_horizon"][0, 0].astype("int64")
-        actions_matlab = mat_contents["actions"].astype("int64") - 1
-        qs_matlab = mat_contents["qs"][0]
-        xn_matlab = mat_contents["xn"][0]
-        vn_matlab = mat_contents["vn"][0]
-
-        likelihoods_matlab = mat_contents["likelihoods"][0]
-
-        num_obs, num_states, _, num_factors = get_model_dimensions(A, B)
-        obs = convert_observation_array(obs_matlab, num_obs)
-        T = len(obs)
-
-        agent = Agent(A=A, B=B, C=C, inference_algo="MMP", policy_len=1, 
-                        inference_horizon=t_horizon, use_BMA = False, 
-                        policy_sep_prior = True)
-        
-        actions_python = np.zeros(T)
-
-        for t in range(T):
-            o_t = (np.where(obs[t])[0][0],)
-            qx, xn_t, vn_t = agent.infer_states_test(o_t)
-            q_pi, efe= agent.infer_policies()
-            action = agent.sample_action()
-
-            actions_python[t] = action
-
-            xn_python = build_xn_vn_array(xn_t)
-            vn_python = build_xn_vn_array(vn_t)
-
-            if t == T-1:
-                xn_python = xn_python[:,:,:-1,:]
-                vn_python = vn_python[:,:,:-1,:]
-
-            start_tstep = max(0, agent.curr_timestep - agent.inference_horizon)
-            end_tstep = min(agent.curr_timestep + agent.policy_len, T)
-
-            xn_validation = xn_matlab[0][:,:,start_tstep:end_tstep,t,:]
-            vn_validation = vn_matlab[0][:,:,start_tstep:end_tstep,t,:]
-
-            self.assertTrue(np.isclose(xn_python, xn_validation).all())
-            self.assertTrue(np.isclose(vn_python, vn_validation).all())
-        
-        self.assertTrue(np.isclose(actions_matlab[0,:],actions_python[:-1]).all())
-
-    def test_active_inference_SPM_1b(self):
-       """
-        Test against output of SPM_MDP_VB_X.m
-        1B - one hidden state factor, one observation modality, backwards horizon = 3, policy_len = 1, policy-conditional prior
-        """
-        
+        self.assertEqual(len(agent.prev_obs), T)
+        self.assertEqual(len(agent.prev_actions), T)
 
 if __name__ == "__main__":
     unittest.main()
+
+
