+++ test/test_inference.py
@@ -1,118 +1,114 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+""" Unit Tests
+__author__: Conor Heins, Alexander Tschantz, Daphne Demekas, Brennan Klein
+"""
+
 import os
 import unittest
 
 import numpy as np
-from scipy.io import loadmat
 
-from pymdp import core
-from pymdp.core.utils import (
-    get_model_dimensions,
-    random_A_matrix,
-    random_B_matrix,
-    obj_array,
-    onehot,
-    norm_dist,
-)
-from pymdp.core.maths import get_joint_likelihood_seq
-from pymdp.core.inference import update_posterior_states_v2, average_states_over_policies
-from pymdp.core.control import update_posterior_policies_mmp
-from pymdp.core import utils
+from pymdp.core import utils, maths
+from pymdp.core import inference
 
+class TestInference(unittest.TestCase):
 
-def rand_onehot_obs(num_obs):
-    if type(num_obs) is int:
-        num_obs = [num_obs]
-    obs = obj_array(len(num_obs))
-    for i in range(len(num_obs)):
-        ob = np.random.randint(num_obs[i])
-        obs[i] = onehot(ob, num_obs[i])
-    return obs
-
-
-def rand_controls(num_controls):
-    if type(num_controls) is int:
-        num_controls = [num_controls]
-    controls = np.zeros(len(num_controls))
-    for i in range(len(num_controls)):
-        controls[i] = np.random.randint(num_controls[i])
-    return controls
-
-
-def rand_dist_states(num_states):
-    if type(num_states) is int:
-        num_states = [num_states]
-    states = obj_array(len(num_states))
-    for i in range(len(num_states)):
-        states[i] = norm_dist(np.random.rand(num_states[i]))
-    return states
-
-
-class Inference(unittest.TestCase):
-    def test_update_posterior_states_v2(self):
+    def test_update_posterior_states(self):
         """
-        Testing our SPM-ified version of `run_MMP` with
-            1 hidden state factor & 1 outcome modality, at a random fixed
-            timestep during the generative process
+        Tests the refactored version of `update_posterior_states`
         """
 
-        past_len = 3
-        future_len = 4
-        num_policies = 5
-        num_states = [6, 7, 8]
-        num_controls = [9, 10, 11]
-        num_obs = [12, 13, 14]
-        num_modalities = len(num_obs)
+        '''Test with single hidden state factor and single observation modality'''
 
-        A = random_A_matrix(num_obs, num_states)
-        B = random_B_matrix(num_states, num_controls)
-        prev_obs = [rand_onehot_obs(num_obs) for _ in range(past_len)]
-        prev_actions = np.array([rand_controls(num_controls) for _ in range(past_len)])
-        policies = [
-            np.array([rand_controls(num_controls) for _ in range(future_len)])
-            for _ in range(num_policies)
-        ]
-        prior = rand_dist_states(num_states)
+        num_states = [3]
+        num_obs = [3]       
 
-        qs_seq_pi, VFE_policies = update_posterior_states_v2(A, B, prev_obs, policies, prev_actions, prior=prior, policy_sep_prior = False)
+        prior = utils.random_single_categorical(num_states)
 
-        qs_seq_pi_future = utils.obj_array(num_policies)
-        for p_idx in range(num_policies):
-            qs_seq_pi_future[p_idx] = qs_seq_pi[p_idx][(1 + past_len) :]
-        
-        # create  C matrix
-        # horizon = len(qs_seq_pi_future[0])
-        # C = utils.obj_array(horizon)
-        # for t in range(horizon):
-        #     C[t] = utils.obj_array(num_modalities)
-        #     for g in range(num_modalities):
-        #         C[t][g] = np.ones(num_obs[g]) 
-        
-        C = utils.obj_array_uniform(num_obs)
-        
-        q_pi, efe = update_posterior_policies_mmp(
-            qs_seq_pi_future,
-            A,
-            B,
-            C,
-            policies,
-            use_utility=True,
-            use_states_info_gain=True,
-            use_param_info_gain=False,
-            prior = None,
-            pA=None,
-            pB=None,
-            F = VFE_policies,
-            E = None,
-            gamma=16.0,
-            return_numpy=True,
-        )
+        A = utils.to_arr_of_arr(maths.softmax(np.eye(num_states[0]) * 0.1))
 
-        qs_pi_curr_t = utils.obj_array(num_policies)
-        for p_idx in range(num_policies):
-            qs_pi_curr_t[p_idx] = qs_seq_pi[p_idx][past_len]
+        obs_idx = 1
+        obs = utils.onehot(obs_idx, num_obs[0])
+
+        qs_out = inference.update_posterior_states(A, obs, prior=prior)
+        qs_validation =  maths.softmax(maths.spm_log_single(A[0][obs_idx,:]) + maths.spm_log_single(prior[0]))
+
+        self.assertTrue(np.isclose(qs_validation, qs_out[0]).all())
+
+        '''Try single modality inference where the observation is passed in as an int'''
+        qs_out_2 = inference.update_posterior_states(A, obs_idx, prior=prior)
+        self.assertTrue(np.isclose(qs_out_2[0], qs_out[0]).all())
+
+        '''Try single modality inference where the observation is a one-hot stored in an object array'''
+        qs_out_3 = inference.update_posterior_states(A, utils.to_arr_of_arr(obs), prior=prior)
+        self.assertTrue(np.isclose(qs_out_3[0], qs_out[0]).all())
+
+        '''Test with multiple hidden state factors and single observation modality'''
+
+        num_states = [3, 4]
+        num_obs = [3]
+
+        prior = utils.random_single_categorical(num_states)
+
+        A = utils.random_A_matrix(num_obs, num_states)
+
+        obs_idx = 1
+        obs = utils.onehot(obs_idx, num_obs[0])
+
+        qs_out = inference.update_posterior_states(A, obs, prior=prior, num_iter = 1)
+
+        # validate with a quick n' dirty implementation of FPI
+
+        # initialize posterior and log prior
+        qs_valid_init = utils.obj_array_uniform(num_states)
+        log_prior = maths.spm_log_obj_array(prior)
+
+        qs_valid_final = utils.obj_array(len(num_states))
+
+        log_likelihood = maths.spm_log_single(maths.get_joint_likelihood(A, obs, num_states))
+
+        num_factors = len(num_states)
+
+        qs_valid_init_all = qs_valid_init[0]
+        for factor in range(num_factors-1):
+            qs_valid_init_all = qs_valid_init_all[...,None]*qs_valid_init[factor+1]
+        LL_tensor = log_likelihood * qs_valid_init_all
+
+        factor_ids = range(num_factors)
+
+        for factor, qs_f in enumerate(qs_valid_init):
+            ax2sum = tuple(set(factor_ids) - set([factor])) # which axes to sum out
+            qL = LL_tensor.sum(axis = ax2sum) / qs_f
+            qs_valid_final[factor] = maths.softmax(qL + log_prior[factor])
+
+        for factor, qs_f_valid in enumerate(qs_valid_final):
+            self.assertTrue(np.isclose(qs_f_valid, qs_out[factor]).all())
+
+        '''Test with multiple hidden state factors and multiple observation modalities, for two different kinds of observation input formats'''
         
-        qs_bma = average_states_over_policies(qs_pi_curr_t, q_pi) # Bayesian model average of hidden states across policies
+        num_states = [3, 4]
+        num_obs = [3, 3, 5]
+
+        prior = utils.random_single_categorical(num_states)
+
+        A = utils.random_A_matrix(num_obs, num_states)
+
+        obs_index_tuple = tuple([np.random.randint(obs_dim) for obs_dim in num_obs])
+
+        qs_out1 = inference.update_posterior_states(A, obs_index_tuple, prior=prior)
+
+        obs_onehots = utils.obj_array(len(num_obs))
+        for g in range(len(num_obs)):
+            obs_onehots[g] = utils.onehot(obs_index_tuple[g], num_obs[g])
+
+        qs_out2 = inference.update_posterior_states(A, obs_onehots, prior=prior)
+
+        for factor in range(len(num_states)):
+            self.assertTrue(np.isclose(qs_out1[factor], qs_out2[factor]).all())
+
+    
 
 if __name__ == "__main__":
     unittest.main()
-
